dummy_client.py
from flask import Flask, redirect, request, session
import urllib.parse
import os
from dotenv import load_dotenv


load_dotenv()


app = Flask(__name__)

app.secret_key = os.getenv("FLASK_SECRET_KEY")


CLIENT_ID = os.getenv("SPOTIPY_CLIENT_ID")
CLIENT_SECRET = os.getenv("SPOTIPY_CLIENT_SECRET")
# This will later be your ngrok HTTPS URL + /callback
REDIRECT_URI = "https://b3ea-212-75-103-72.ngrok-free.app/callback"
AUTH_URL = "https://accounts.spotify.com/authorize"
TOKEN_URL = "https://accounts.spotify.com/api/token"


@app.route("/login")
def login():
    params = {
        "client_id": CLIENT_ID,
        "response_type": "code",
        "redirect_uri": REDIRECT_URI,
        "scope": "user-read-private user-read-email",
    }
    return redirect(f"{AUTH_URL}?{urllib.parse.urlencode(params)}")


@app.route("/callback")
def callback():
    code = request.args.get("code")
    return f"Code: {code}"


if __name__ == "__main__":
    # just HTTP on localhost:5000
    app.run(host="127.0.0.1", port=5000)
