requirements.txt
spotipy
python-dotenv
fastapi
uvicorn
python-multipart
Flask
spotipy
openai
pre-commit
types-requests
